requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.1.0
cachetools==5.3.0
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-webpack-loader==1.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.2
msgpack==1.0.5
mysql-connector-python==8.1.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psycopg2-binary==2.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.4.0
ytmusicapi==1.0.0
zope.interface==6.1
